Move User filter into a separate module
