+++ app/filters/activity.py
@@ -12,7 +12,7 @@ from app.filters.common import (
     IdFilterMixin,
 )
 from app.filters.entity import NestedEntityFilter
-from app.filters.person import CreatorFilterMixin
+from app.filters.user import CreatorFilterMixin
 
 NestedUsedFilterDep = FilterDepends(with_prefix("used", NestedEntityFilter))
 NestedGeneratedFilterDep = FilterDepends(with_prefix("generated", NestedEntityFilter))
