+++ app/filters/entity.py
@@ -8,7 +8,7 @@ from app.db.types import DerivationType, EntityLifecycleStatus, EntityType
 from app.dependencies.filter import FilterDepends
 from app.filters.base import CustomFilter
 from app.filters.common import AuthorizedFilterMixin, CreationFilterMixin, IdFilterMixin
-from app.filters.person import CreatorFilterMixin
+from app.filters.user import CreatorFilterMixin
 
 
 class NestedDerivationFilter(CustomFilter):
