+++ app/filters/external_url.py
@@ -12,7 +12,7 @@ from app.filters.common import (
     ILikeSearchFilterMixin,
     NameFilterMixin,
 )
-from app.filters.person import CreatorFilterMixin
+from app.filters.user import CreatorFilterMixin
 
 
 class NestedExternalUrlFilter(
