+++ app/filters/measurement_label.py
@@ -11,7 +11,7 @@ from app.filters.common import (
     IdFilterMixin,
     PrefLabelMixin,
 )
-from app.filters.person import CreatorFilterMixin
+from app.filters.user import CreatorFilterMixin
 
 
 class NestedMeasurementLabelFilter(
