+++ app/filters/scientific_artifact_external_url_link.py
@@ -5,11 +5,11 @@ from app.dependencies.filter import FilterDepends
 from app.filters.base import CustomFilter
 from app.filters.common import IdFilterMixin
 from app.filters.external_url import NestedExternalUrlFilter, NestedExternalUrlFilterDep
-from app.filters.person import CreatorFilterMixin
 from app.filters.scientific_artifact import (
     NestedScientificArtifactFilter,
     NestedScientificArtifactFilterDep,
 )
+from app.filters.user import CreatorFilterMixin
 
 
 class ScientificArtifactExternalUrlLinkFilter(
