+++ app/filters/scientific_artifact_publication_link.py
@@ -5,12 +5,12 @@ from app.db.types import PublicationType
 from app.dependencies.filter import FilterDepends
 from app.filters.base import CustomFilter
 from app.filters.common import IdFilterMixin
-from app.filters.person import CreatorFilterMixin
 from app.filters.publication import NestedPublicationFilter, NestedPublicationFilterDep
 from app.filters.scientific_artifact import (
     NestedScientificArtifactFilter,
     NestedScientificArtifactFilterDep,
 )
+from app.filters.user import CreatorFilterMixin
 
 
 class ScientificArtifactPublicationLinkFilter(
